REST-API/Python/requirements.txt
aniso8601==1.1.0
arrow==0.7.0
Flask==0.10.1
Flask-HTTPAuth==2.7.1
Flask-RESTful==0.3.5
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
passlib==1.6.5
python-dateutil==2.4.2
pytz==2015.7
six==1.10.0
SQLAlchemy==1.0.11
Werkzeug==0.11.3
